requirements.txt

validators
pandas
python-decouple
requests
beautifulsoup4
pathlib
SQLAlchemy
